# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://10.66.4.47:8989",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/staff/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get all services",
                "description": "Returns a list of all services. Admin only.",
                "operationId": "331c1b759033931039c1bd2bb04e0c6b",
                "responses": {
                    "200": {
                        "description": "List of all services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Haircut"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Basic haircut service",
                                                "nullable": true
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "duration": {
                                                "type": "integer",
                                                "example": 30
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 25
                                            },
                                            "is_active": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "preparation_time": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "cleanup_time": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create a new service",
                "operationId": "e61b20d48778959a3be61a61aee3cc9c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServiceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service created"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/staff/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get service by ID",
                "operationId": "0bba1fe0d0fe0b8e1275548b8a47033a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Haircut"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Basic haircut service",
                                            "nullable": true
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "duration": {
                                            "type": "integer",
                                            "example": 30
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 25
                                        },
                                        "is_active": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "preparation_time": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "cleanup_time": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Update a service",
                "operationId": "df8237c0505ea183cb9085614060c643",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServiceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service updated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete a service",
                "operationId": "efff6ed2301055ed51d5d61a929caced",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Service deleted"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Service not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ServiceRequest": {
                "required": [
                    "name",
                    "category_id",
                    "duration",
                    "price",
                    "preparation_time",
                    "cleanup_time"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Haircut"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 5000,
                        "example": "Basic haircut service",
                        "nullable": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "duration": {
                        "type": "integer",
                        "minimum": 5,
                        "example": 30
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "minimum": 0.01,
                        "example": 25
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "preparation_time": {
                        "type": "integer",
                        "minimum": 0,
                        "example": 5
                    },
                    "cleanup_time": {
                        "type": "integer",
                        "minimum": 0,
                        "example": 5
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}